+++ dsmrreader/config/base.py
@@ -154,15 +154,16 @@ LANGUAGES = (
 LOCALE_PATHS = (os.path.join(BASE_DIR, 'locales'), )
 
 
-""" DSMR Project settings. """
+""" DSMR-reader Project settings. """
 
-DSMR_SUPPORTED_DB_VENDORS = ('postgresql', 'mysql')
+DSMRREADER_SUPPORTED_DB_VENDORS = ('postgresql',)  # In the past, 'mysql' was supported as well.
 
-DSMR_BACKUP_DIRECTORY = 'backups'  # Relative to project root.
-DSMR_DROPBOX_SYNC_INTERVAL = 1  # Only check for changes once per hour.
+DSMRREADER_BACKUP_DIRECTORY = 'backups'  # Relative to project root.
+DSMRREADER_DROPBOX_SYNC_INTERVAL = 1  # Only check for changes once per hour.
 
-DSMR_MANAGEMENT_COMMANDS_PID_FOLDER = '/var/tmp/'
+DSMRREADER_MANAGEMENT_COMMANDS_PID_FOLDER = '/var/tmp/'
 
 DSMRREADER_VERSION = dsmrreader.__version__
-DSMR_RAW_VERSION = dsmrreader.VERSION
-DSMR_LATEST_VERSION_FILE = 'https://raw.githubusercontent.com/dennissiemensma/dsmr-reader/master/dsmrreader/__init__.py'
+DSMRREADER_RAW_VERSION = dsmrreader.VERSION
+DSMRREADER_LATEST_VERSION_FILE = 'https://raw.githubusercontent.com/dennissiemensma/dsmr-reader/' \
+    'master/dsmrreader/__init__.py'